grpc-server-microservice/requirements.txt
grpcio
grpcio-tools
python-dotenv
pytest-grpc
peewee
psycopg2-binary